feat: 블랙잭 판정 및 1.5배 수익 계산 기능 구현
- Hand에 isBlackjack 추가 (2장이면서 21점인 경우 판별)
- Participant에 isBlackjack 메서드 생성(딜러, 플레이어 둘 다 사용)
- GameResult에 BLACKJACK(1.5배) enum 값 추가
- GameResults.create에서 블랙잭 우선 판정 후 일반 승패 비교(양쪽 모두 블랙잭이면 무승부, 플레이어만 블랙잭이면 1.5배 수익)
- GameResult.reverse에 BLACKJACK 케이스(플레이어만 블랙잭)의 반대(딜러의 승패)는 무조건 패
- OutputView에 BLACKJACK 결과는 WIN으로 변경
- Money.multiply를 double로 변경하여 블랙잭시 1.5배 연산 지원

+++ src/main/java/blackjack/domain/money/Money.java
@@ -14,8 +14,8 @@ public class Money {
         return new Money(value + money.value);
     }
 
-    public Money multiply(int multiplier) {
-        return new Money(value * multiplier);
+    public Money multiply(double multiplier) {
+        return new Money((int) (value * multiplier));
     }
 
     public Money negate() {
